pyproject.toml
[project]
name = "p2p-steam"
version = "0.1.0"
description = "P2P Steam Platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi[all]>=0.115.6",
    "sqlalchemy>=2.0.37",
    "pyjwt>=2.10.1",
]

[dependency-groups]
dev = [
    "alembic>=1.14.0",
    "coverage>=7.6.10",
    "pytest-asyncio==0.20.3",
    "pytest==7.2.1",
    "pytest-mock>=3.14.0",
    "pytest-xdist>=3.6.1",
    "ruff>=0.9.1",
    "faker>=33.3.1",
    "greenlet>=3.1.1"
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = "tests"
addopts = [
    "--import-mode=importlib"
]
asyncio_mode = "auto"

[tool.ruff.lint]
extend-select = ["E"]

[tool.ruff.lint.per-file-ignores]
"env.py" = ["F401", "F811"]
"src/users/models/user_model.py" = ["F821"]
"src/accounts/models/account_model.py" = ["F821"]
"src/transactions/models/transaction_model.py" = ["F821"]